Origin: move out of `internal`
As origin is now saved as part of quasiquotes, let's make sure it abides
by MIMA standards (which are not currently applied to `internal` code).

Similarly, remove TokenStreamPosition and inline in Origin.Parsed.
