+++ scalameta/parsers/shared/src/main/scala/scala/meta/internal/parsers/Pos.scala
@@ -1,8 +1,8 @@
 package scala.meta.internal.parsers
 
 import scala.meta.Tree
-import scala.meta.internal.trees.Origin
 import scala.meta.prettyprinters._
+import scala.meta.trees.Origin
 
 // NOTE: `startTokenPos` and `endTokenPos` are BOTH INCLUSIVE.
 // This is at odds with the rest of scala.meta, where ends are non-inclusive.
